Rename modules, change downloader import
